started upgrading the app
