+++ frontend/package.json
@@ -32,7 +32,6 @@
     "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3",
     "styled-components": "^5.3.3",
-    "typescript": "^4.8.4",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
@@ -40,12 +39,6 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
   },
   "browserslist": {
     "production": [
@@ -63,7 +56,8 @@
     "@types/react-dom": "^18.0.6",
     "@types/react-router-dom": "^5.3.3",
     "devextreme-themebuilder": "^21.1.5",
-    "nodemon": "^2.0.14"
+    "nodemon": "^2.0.14",
+    "typescript": "^4.8.4"
   },
   "proxy": "http://localhost:5000"
 }
