+++ frontend/src/components/signup/signup.js
@@ -45,7 +45,6 @@ export default function SignUp() {
   const handleSubmit = (event) => {
     event.preventDefault();
     const data = new FormData(event.currentTarget);
-    // eslint-disable-next-line no-console
 
     console.log(data.get('userType'));
